move vcr config to global conftest
